requirements.txt
flask==3.1.0
flask-cors==5.0.0
easyocr==1.7.2
python-doctr[torch]==0.10.0
PyMuPDF==1.25.5
Pillow==11.2.1
numpy==1.26.4
opencv-python-headless==4.11.0.86
streamlit>=1.30.0
ultralytics>=8.0.0
pandas>=2.0.0
